"Refactor image arrow.svg and implement the swiper library in ClientsSlider.vue"
This update includes two significant changes. First, the arrow.svg image in the public space was refactored to maintain a consistent visual display across all screens. Second, the 'swiper' library was implemented in ClientsSlider.vue, which resulted in significant changes in the .vue component structure in order to adapt to the swiper's required markup. This library was added to make the client's slider feature more interactive and user-friendly, providing benefits like looping, pagination and navigation.