Add new feature components and update existing ones for Customer and Home Pages
This update introduces new feature components to the Customer and Home page, and makes necessary changes to the existing components to improve the UI/UX and present the information in a more organized manner.

The diff shows changes to the Customers.vue file which suggests additional icons have been included as part of the compareItems objects. This would enable users to differentiate and compare items more efficiently. Moreover, changes in the Header.vue file suggest that visual representation for services section has been improved. Also, updates in the InterestChannelsBlock.vue page improved the structure and presentation of the text there. The 'main', 'integr', and 'slogan' classes in SloganBlock.vue are updated to add relative positioning and additional padding. The Footer.vue, ComparisonCard.vue, and app.css are updated significantly to enhance the look and feel.

These changes are made to improve the overall user experience when navigating through these pages, and to make the information easily accessible and path to action more distinct for visitors.