Remove Dashboard.vue, Add ChannelsCatalog, Login Controller and others
The Dashboard.vue was deleted because it is replaced with ChannelsCatalog.vue which is more congruent with the new app structure. The Login.vue was updated with a different route. The new addition is the Header.vue in the Dashboard, containing app's navigation. Another newly added file is the LoginController.php, which contains logic for user authentication. Finally, minor changes were done in CreateNewUser.php and AppLayout.vue.