Adds total balance and personal data to the profile section
This update creates a new template page including total balance and personal data fields with all necessary routing and controllers. These changes are necessary for displaying the total balance and personal data of the user on the profile dashboard. It also adds some interactivity to the header icons such as a click event to the person icon which now redirects to the personal data page. Some improvements were also made to general layouts and design of the profile pages. Another addition is a date localization service for a more versatile date handling.
New styling rules were added to enhance UX.