"Added new routes and methods for pattern manipulations and Refactored Templates component
New routes and methods were added in the web.php and PatternController files respectively, to support operations like renaming, duplicating and removing a pattern. Templates component was refactored to reactively show pattern details and manage a context menu with options for manipulation. 'UpdatePatternRequest' was also slightly changed to allow a nullable 'body'.  'AddingTemplate' component now effectively count the created patterns. A title field was added to the patterns table. The changes were needed to enhance pattern manipulation and provide richer user interaction with the application.

The DateLocalizationService was also slightly refactored to improve date handling."