Add default title attribute to Pattern model
In this commit, we've moved the default value for the `title` attribute in the Pattern model from migration to the model itself. This change is done because it provides a cleaner and more organized code structure. Also, it makes easier to handle the model defaults in the model class and lets us avoid having defaults spread across migrations. This approach is more maintainable for future codebase development.