"Enhance chat server and improve base UI components"
Updated the ChatWebSocketServer to parse users' unique ID and relay messages to a particular user for enhanced targeting. Improved WebSocket connection in the Chat Support feature by appending the user's ID. Introduced echo messages after sending a message to the moderator for better debugging and user experience.

Enhanced UI components in various places. Added CardBox in the Admin's SupportChatView and improved the look of cards in the Support page. Modified Admin's Tables to show messages corresponding to the user's ID. Made Messenger respond corresponding to messages and incoming events.

Made style adjustments and improvements for responsive design in support, messenger, and ticket cards. Improved the handling of WebSocket connections and messages in the Dashboard and Admin pages. Also, changed some labels in the Admin's TableChannels component for clarity.

This commit aims for more personalized chat experience and enhanced user interface usability and response on different platforms.

Took 1 hour 47 minutes