Implement multiple images upload functionality
The changes made allow for multiple images to be uploaded at once, with a restriction of a maximum of 10. Added functionalities to delete previously uploaded images in the editor, and updated the method of capturing and sending image data through form to support multiple images. Code has also been adjusted to warn users if they exceed the upload limit. Also, code related to handling the display of image previews in the editor and post-preview sections was updated.

Took 2 hours 18 minutes