Refactor code and update logic for orders and channels
The commit includes multiple changes like using 'if-elseif' instead of switch statement for improved readability in ChannelService and updates to the logic for handling orders in OrderController and OrderService. It also includes addition of new 'status' options in migrations for the orders table and changes related to handling channels in ChannelController. Additionally, the front-end was updated for better error handling and implementing i18n for multilinguism.

Took 5 hours 54 minutes