Adjust schedule command for updating finished orders
The Kernel.php file has been updated to incorporate the use of 'UpdateFinishedOrdersCommand' class in the scheduling of tasks. Instead of directly referencing the command string 'update:finished-orders', we are now using the class method to call this command and it is now set to run in the background on an hourly schedule. This adjustment enhances the code organization and ensures better resource management by running the command in the background.

Took 5 minutes